Make a release for CentOS 7
This replaces the release stage with a deploy step for each existing
job.  As before, the release is only created for tags in the main
repository.  However, unlike before, two releases are uploaded, one for
each job ("debian-10" and "centos-7").

For that to work, the release is compiled in Docker instead of directly
in the Travis VM.