Fix certificate check with --no-check-certificate
This fixes the bypass for the certificate checking by also disabling
host verification (not just the signature).

Note that the certificate of the proxy, if it is accessed via HTTPS, is
always checked.  See the following for more information:

- https://curl.se/libcurl/c/CURLOPT_PROXY_SSL_VERIFYHOST.html
- https://curl.se/libcurl/c/CURLOPT_PROXY_SSL_VERIFYPEER.html

I didn't find how to fix that with ocurl so that's a limitation of our
client for now.