Fix upload of large files to MinIO
This uses the `CURLFORM_FILE` option instead of `CURLFORM_FILECONTENT`.
This makes the field a real file upload field.  Indeed, according to the
documentation of `CURLFORM_FILECONTENT`
(https://curl.se/libcurl/c/curl_formadd.html#CURLFORMFILECONTENT):

> This part does not automatically become a file upload part simply
> because its data was read from a file.

The change is visible in the multipart form data.  We now have:

    Content-Disposition: form-data; name="file"; filename="trace.cst.gz"
    Content-Type: application/octet-stream

instead of just:

    Content-Disposition: form-data; name="file"

I think this makes MinIO treat this part as the uploaded file, which
avoids the file size error.

I managed to achieve the same result with the `curl_mime_*` functions,
which are supposed to replace `curl_form*`.  Unfortunately, this new API
is only available for libcurl 7.56 and above, which makes it unsuitable
for our use case where CentOS 7, with libcurl 7.29, is a common
environment.