Fix YoJson error when uploading a trace
The following error was raised when uploading a trace:

```
cs-api: internal error, uncaught exception:
        Yojson__Basic.Util.Type_error("Expected int or null, got string", _)
```

It turned out that the api uses a string in form data for the
`success_action_status` field, while cs-api was expecting to parse an
int. This commit fixes this by expecting a string in cs-api.