Use Opam 2.2 on Windows (via setup-ocaml@v3)
This upgrades the OCaml action used in CI, so that we use the same Opam
as for Linux and get the same benefits.

The only downside is that we have to fix some package definitions, hence
the custom `opam_win_repo` repository. Hopefully this is temporary and
won't be necessary for long (once the official repository will have
caught up).