Release version 2.7.1
Changes for the user:

- The Windows release ships with a CA certificates bundle.

Comments:

This is needed because on Windows we can't rely on Schannel and seem to
be using OpenSSL instead. To help the user get started we provide a CA
bundle they can use to get started. They will probably need to adjust it
if they have custom certificates.

In the future, it might make sense to try and upgrade `ocurl` to be able
to use more options of libcurl and try to make it use Schannel, and
determine whether we need to compile the DLL ourselves instead of
downloading it from the official website. Currently this is definitely
tricky because we can't use the new ocurl on CentOS 7, and using 8+
would unnecessarily decrease the compatibility of our uploader.